# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AdminEvaluationCorrection": {
        "properties": {
          "evaluacion": {
            "$ref": "#/components/schemas/AdminEvaluationDetail"
          },
          "mutation": {
            "$ref": "#/components/schemas/AdminMutation"
          }
        },
        "required": [
          "mutation",
          "evaluacion"
        ],
        "title": "AdminEvaluationCorrection",
        "type": "object"
      },
      "AdminEvaluationDetail": {
        "description": "Detalle normalizado; mantiene el mismo contrato que una fila de grilla.",
        "properties": {
          "captured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "detalle": {
            "additionalProperties": true,
            "title": "Detalle",
            "type": "object"
          },
          "empresa": {
            "default": "",
            "title": "Empresa",
            "type": "string"
          },
          "evaluador": {
            "default": "",
            "title": "Evaluador",
            "type": "string"
          },
          "evaluador_dni": {
            "default": "",
            "title": "Evaluador Dni",
            "type": "string"
          },
          "evaluador_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluador Id"
          },
          "fecha": {
            "format": "date",
            "title": "Fecha",
            "type": "string"
          },
          "fundo": {
            "default": "",
            "title": "Fundo",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lote": {
            "default": "",
            "title": "Lote",
            "type": "string"
          },
          "lote_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lote Id"
          },
          "module_key": {
            "enum": [
              "estadios",
              "flores",
              "baya",
              "pesos",
              "brotes",
              "ramas"
            ],
            "title": "Module Key",
            "type": "string"
          },
          "modulo": {
            "default": "",
            "title": "Modulo",
            "type": "string"
          },
          "origen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origen"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "source_table": {
            "title": "Source Table",
            "type": "string"
          },
          "tipo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo"
          },
          "variedad": {
            "default": "",
            "title": "Variedad",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_table",
          "source_id",
          "module_key",
          "fecha"
        ],
        "title": "AdminEvaluationDetail",
        "type": "object"
      },
...
      "AdminEvaluationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminEvaluationItem"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "AdminEvaluationPage",
        "type": "object"
      },
...
      "ApiError": {
        "description": "Error legible para la app y para Swagger.",
        "example": {
          "detail": "Descripción del error"
        },
        "properties": {
          "detail": {
            "description": "Motivo por el que la operación no pudo completarse",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageInfo": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API interna para registrar evaluaciones de app-campo en PostgreSQL. Incluye autenticación administrativa y lectura paginada para el panel Angular.",
    "title": "Aqu Anqa · API de campo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/evaluaciones": {
      "get": {
        "description": "Consulta server-side las seis familias de evaluación y aplica filtros por ubicación, evaluador, módulo y fecha. La ruta es de lectura; la autorización RBAC se integra antes de publicar el panel fuera de la red interna.",
        "operationId": "listarEvaluacionesAdmin",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "module_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "estadios",
                    "flores",
                    "baya",
                    "pesos",
                    "brotes",
                    "ramas"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Key"
            }
          },
          {
            "in": "query",
            "name": "empresa_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Empresa Id"
            }
          },
          {
            "in": "query",
            "name": "fundo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fundo Id"
            }
          },
          {
            "in": "query",
            "name": "modulo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modulo Id"
            }
          },
          {
            "in": "query",
            "name": "lote_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lote Id"
            }
          },
          {
            "in": "query",
            "name": "evaluador_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evaluador Id"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "captured_at",
              "enum": [
                "captured_at",
                "fecha",
                "module_key",
                "empresa",
                "fundo",
                "modulo",
                "lote",
                "evaluador"
              ],
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_dir",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Sort Dir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEvaluationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Se requiere un token Bearer"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Se requiere una sesión administrativa válida."
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "El usuario no tiene permiso para abrir el panel administrativo"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La sesión no tiene permiso para el panel administrativo."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo consultar la superficie administrativa"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "PostgreSQL no está disponible o la consulta administrativa falló."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lista evaluaciones para administración",
        "tags": [
          "Administración"
        ]
      }
    },
...
    "/v1/admin/evaluaciones/{module_key}/{source_id}": {
      "get": {
        "operationId": "consultarDetalleEvaluacionAdmin",
        "parameters": [
          {
            "in": "path",
            "name": "module_key",
            "required": true,
            "schema": {
              "enum": [
                "estadios",
                "flores",
                "baya",
                "pesos",
                "brotes",
                "ramas"
              ],
              "title": "Module Key",
              "type": "string"
            }
          },
          {
            "description": "ID de la tabla de origen",
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "description": "ID de la tabla de origen",
              "exclusiveMinimum": 0,
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "description": "Tabla física; obligatorio para distinguir las dos fuentes de baya",
            "in": "query",
            "name": "source_table",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tabla física; obligatorio para distinguir las dos fuentes de baya",
              "title": "Source Table"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEvaluationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Se requiere un token Bearer"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Se requiere una sesión administrativa válida."
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "El usuario no tiene permiso para abrir el panel administrativo"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La sesión no tiene permiso para el panel administrativo."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Evaluación no encontrada"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La evaluación no existe."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo consultar la superficie administrativa"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "PostgreSQL no está disponible o la consulta administrativa falló."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consulta el detalle normalizado de una evaluación",
        "tags": [
          "Administración"
        ]
      },
      "patch": {
        "description": "Valida el payload con las reglas del módulo y aplica una mutación allowlisted en una transacción PostgreSQL. Conserva antes/después, actor, comentario e idempotency_key.",
        "operationId": "corregirEvaluacionAdmin",
        "parameters": [
          {
            "in": "path",
            "name": "module_key",
            "required": true,
            "schema": {
              "enum": [
                "estadios",
                "flores",
                "baya",
                "pesos",
                "brotes",
                "ramas"
              ],
              "title": "Module Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Source Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEvaluationCorrection"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Se requiere un token Bearer"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Se requiere una sesión administrativa válida."
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "El usuario no tiene permiso para abrir el panel administrativo"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La sesión no tiene permiso para el panel administrativo."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "La evaluación no existe o está fuera de alcance"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La evaluación no existe."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "idempotency_key ya fue utilizado para otro cambio"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La operación entra en conflicto."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "payload_invalido": {
                    "value": {
                      "detail": [
                        {
                          "input": {},
                          "loc": [
                            "body",
                            "module_key"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ]
                    }
                  },
                  "regla_negocio": {
                    "value": {
                      "detail": "La corrección requiere valores"
                    }
                  }
                },
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "Los valores no cumplen las reglas del módulo."
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo consultar la superficie administrativa"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "PostgreSQL no está disponible o la consulta administrativa falló."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Corrige una evaluación individual",
        "tags": [
          "Administración"
        ]
      }
    },
...
    "/v1/evaluaciones": {
      "post": {
        "description": "Operación idempotente por client_id. El primer envío devuelve 201; un reenvío idéntico devuelve 200 y status=duplicate; reutilizar el UUID con otro contenido devuelve 409.",
        "operationId": "registrarEvaluacion",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "captura_estadios": {
                  "summary": "Evaluación de estadios creada por Flutter",
                  "value": {
                    "cortina": 1,
                    "evaluador_dni": "10616663",
                    "fecha": "2026-08-31",
                    "hilera": 2,
                    "id": "11111111-1111-4111-8111-111111111111",
                    "lote_id": 12,
                    "module_key": "estadios",
                    "planta": 3,
                    "valores": {
                      "m1_e1": 1,
                      "m1_e2": 2,
                      "m1_e3": 3,
                      "m1_e4": 4,
                      "m1_e5": 5
                    }
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/EvaluationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "client_id ya fue utilizado con otro módulo o contenido"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "El client_id ya fue usado con contenido diferente."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "payload_invalido": {
                    "value": {
                      "detail": [
                        {
                          "input": {},
                          "loc": [
                            "body",
                            "module_key"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ]
                    }
                  },
                  "regla_negocio": {
                    "value": {
                      "detail": "lote_id no existe o no es elegible para captura"
                    }
                  }
                },
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "La captura o sus referencias de maestro no son válidas."
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo guardar la evaluación en PostgreSQL"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "PostgreSQL no está disponible o la transacción falló."
          }
        },
        "summary": "Registra una evaluación y actualiza core",
        "tags": [
          "Evaluaciones"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Servidor actual (local)",
      "url": "/"
    }
  ],
...
}